Finally got everything working right no the wizard. The only change I had to make was adding a wrapper for changing button text.
Moved the last callback (amiRNA4) into one callback, for both filtered and unfiltered. Added a closuer to
handle the click.

Began work on the submission page. AJAX is properly configured, needs to handle the response and parse the
token out of the server response. The DesignerAction also needs to have it's input to the Daemon updated. Validation has already been updated.